Review: Extract TestSiteResourceConfigurationElement
